+++ geode/random/permute.cpp
@@ -19,8 +19,8 @@ static inline uint64_t fe2_encrypt(const int bits, const uint128_t key, const ui
   assert(bits==64 || x<(uint64_t(1)<<bits));
   // Prepare for FE2
   const int a = bits>>1, b = bits-a; // logs of a,b in the paper
-  const uint32_t ma = (uint64_t(1)<<a)-1, // bit masks
-                 mb = (uint64_t(1)<<b)-1;
+  const uint32_t ma = (uint32_t)(uint64_t(1)<<a)-1, // bit masks
+                 mb = (uint32_t)(uint64_t(1)<<b)-1;
   uint32_t L = uint32_t(x>>b), R = x&mb;
   // Three rounds of FE2
   L = ma&(L+cast_uint128<uint32_t>(threefry(key,uint64_t(1)<<32|R))); // round 1: s = a
@@ -34,8 +34,8 @@ static inline uint64_t fe2_decrypt(const int bits, const uint128_t key, const ui
   assert(bits==64 || x<(uint64_t(1)<<bits));
   // Prepare for FE2
   const int a = bits>>1, b = bits-a; // logs of a,b in the paper
-  const uint32_t ma = (uint64_t(1)<<a)-1, // bit masks
-                 mb = (uint64_t(1)<<b)-1;
+  const uint32_t ma = (uint32_t)(uint64_t(1)<<a)-1, // bit masks
+                 mb = (uint32_t)(uint64_t(1)<<b)-1;
   uint32_t L = uint32_t(x>>b), R = x&mb;
   // Three rounds of FE2
   L = ma&(L-cast_uint128<uint32_t>(threefry(key,uint64_t(3)<<32|R))); // round 3: s = a
